+++ 35-semi/35-semi/View/BaseViewController.swift
@@ -1,0 +1,18 @@
+//
+//  BaseViewController.swift
+//  35-semi
+//
+//  Created by 조성민 on 10/15/24.
+//
+
+import UIKit
+
+class BaseViewController: UIViewController {
+    
+    func setStyle() {}
+    
+    func setUI() {}
+    
+    func setLayout() {}
+    
+}
